# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Full Stack FastAPI Project", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/items/": {
      "get": {
        "tags": ["items"],
        "summary": "Read Items",
        "description": "Retrieve items.",
        "operationId": "items-read_items",
        "security": [{ "SupabaseBearer": [] }],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ItemsPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["items"],
        "summary": "Create Item",
        "description": "Create new item.",
        "operationId": "items-create_item",
        "security": [{ "SupabaseBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ItemCreate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ItemPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{id}": {
      "get": {
        "tags": ["items"],
        "summary": "Read Item",
        "description": "Get item by ID.",
        "operationId": "items-read_item",
        "security": [{ "SupabaseBearer": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ItemPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["items"],
        "summary": "Update Item",
        "description": "Update an item.",
        "operationId": "items-update_item",
        "security": [{ "SupabaseBearer": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ItemUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ItemPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["items"],
        "summary": "Delete Item",
        "description": "Delete an item.",
        "operationId": "items-delete_item",
        "security": [{ "SupabaseBearer": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Message" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/providers/": {
      "get": {
        "tags": ["providers"],
        "summary": "Read Providers",
        "description": "Retrieve providers.",
        "operationId": "providers-read_providers",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProvidersPublic" } }
            }
          }
        }
      },
      "post": {
        "tags": ["providers"],
        "summary": "Create Provider Endpoint",
        "description": "Create new provider.",
        "operationId": "providers-create_provider_endpoint",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ProviderCreate" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProviderPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/providers/{id}": {
      "get": {
        "tags": ["providers"],
        "summary": "Read Provider",
        "description": "Get provider by ID.",
        "operationId": "providers-read_provider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProviderPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["providers"],
        "summary": "Update Provider",
        "description": "Update a provider.",
        "operationId": "providers-update_provider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ProviderUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProviderPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["providers"],
        "summary": "Delete Provider Endpoint",
        "description": "Delete a provider.",
        "operationId": "providers-delete_provider_endpoint",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { "type": "string" },
                  "title": "Response Providers-Delete Provider Endpoint"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/posts/": {
      "get": {
        "tags": ["posts"],
        "summary": "Read Posts",
        "operationId": "posts-read_posts",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": { "$ref": "#/components/schemas/PostPublic" },
                  "type": "array",
                  "title": "Response Posts-Read Posts"
                }
              }
            }
          }
        },
        "security": [{ "SupabaseBearer": [] }]
      },
      "post": {
        "tags": ["posts"],
        "summary": "Create Post",
        "description": "Create a new post with optional image upload.\n\nThe image will be uploaded to Supabase Storage and the post will be created\nwith the image URL. The user_id is automatically set from the authenticated user.\n\nFile uploads are handled by the centralized FileUploadService for consistent\nvalidation, error handling, and storage management.",
        "operationId": "posts-create_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/Body_posts-create_post" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PostPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "SupabaseBearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_posts-create_post": {
        "properties": {
          "title": { "type": "string", "maxLength": 255, "minLength": 1, "title": "Title" },
          "content": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Content" },
          "category_id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Category Id"
          },
          "image": {
            "anyOf": [{ "type": "string", "format": "binary" }, { "type": "null" }],
            "title": "Image"
          }
        },
        "type": "object",
        "required": ["title"],
        "title": "Body_posts-create_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ItemCreate": {
        "properties": {
          "title": { "type": "string", "maxLength": 255, "minLength": 1, "title": "Title" },
          "description": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Description"
          }
        },
        "type": "object",
        "required": ["title"],
        "title": "ItemCreate"
      },
      "ItemPublic": {
        "properties": {
          "title": { "type": "string", "maxLength": 255, "minLength": 1, "title": "Title" },
          "description": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Description"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "owner_id": { "type": "string", "format": "uuid", "title": "Owner Id" }
        },
        "type": "object",
        "required": ["title", "id", "owner_id"],
        "title": "ItemPublic"
      },
      "ItemUpdate": {
        "properties": {
          "title": {
            "anyOf": [{ "type": "string", "maxLength": 255, "minLength": 1 }, { "type": "null" }],
            "title": "Title"
          },
          "description": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "ItemUpdate"
      },
      "ItemsPublic": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/ItemPublic" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "ItemsPublic"
      },
      "Message": {
        "properties": { "message": { "type": "string", "title": "Message" } },
        "type": "object",
        "required": ["message"],
        "title": "Message"
      },
      "PostPublic": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "user_id": { "type": "string", "format": "uuid", "title": "User Id" },
          "category_id": {
            "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }],
            "title": "Category Id"
          },
          "title": { "type": "string", "title": "Title" },
          "content": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Content" },
          "image_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Image Url"
          },
          "created_at": { "type": "string", "format": "date-time", "title": "Created At" },
          "updated_at": { "type": "string", "format": "date-time", "title": "Updated At" },
          "category_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Category Name"
          },
          "author": { "anyOf": [{ "$ref": "#/components/schemas/UserPublic" }, { "type": "null" }] }
        },
        "type": "object",
        "required": ["id", "user_id", "title", "created_at", "updated_at"],
        "title": "PostPublic"
      },
...
      "ProviderCreate": {
        "properties": {
          "public_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Public Name"
          },
          "email": { "type": "string", "maxLength": 255, "title": "Email" },
          "accepts_online_bookings": {
            "type": "boolean",
            "title": "Accepts Online Bookings",
            "default": false
          },
          "services": {
            "additionalProperties": { "items": { "type": "string" }, "type": "array" },
            "type": "object",
            "title": "Services"
          },
          "photo_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Photo Url"
          },
          "working_hours": {
            "additionalProperties": { "type": "string" },
            "type": "object",
            "title": "Working Hours"
          }
        },
        "type": "object",
        "required": ["public_name", "email"],
        "title": "ProviderCreate"
      },
      "ProviderPublic": {
        "properties": {
          "public_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Public Name"
          },
          "email": { "type": "string", "maxLength": 255, "title": "Email" },
          "accepts_online_bookings": {
            "type": "boolean",
            "title": "Accepts Online Bookings",
            "default": false
          },
          "services": {
            "additionalProperties": { "items": { "type": "string" }, "type": "array" },
            "type": "object",
            "title": "Services"
          },
          "photo_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Photo Url"
          },
          "working_hours": {
            "additionalProperties": { "type": "string" },
            "type": "object",
            "title": "Working Hours"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" }
        },
        "type": "object",
        "required": ["public_name", "email", "id"],
        "title": "ProviderPublic"
      },
      "ProviderUpdate": {
        "properties": {
          "public_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Public Name"
          },
          "email": { "type": "string", "maxLength": 255, "title": "Email" },
          "accepts_online_bookings": {
            "type": "boolean",
            "title": "Accepts Online Bookings",
            "default": false
          },
          "services": {
            "additionalProperties": { "items": { "type": "string" }, "type": "array" },
            "type": "object",
            "title": "Services"
          },
          "photo_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Photo Url"
          },
          "working_hours": {
            "additionalProperties": { "type": "string" },
            "type": "object",
            "title": "Working Hours"
          }
        },
        "type": "object",
        "title": "ProviderUpdate"
      },
      "ProvidersPublic": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/ProviderPublic" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "ProvidersPublic"
      },
...
      "UserPublic": {
        "properties": {
          "email": { "type": "string", "maxLength": 255, "format": "email", "title": "Email" },
          "is_active": { "type": "boolean", "title": "Is Active", "default": true },
          "is_superuser": { "type": "boolean", "title": "Is Superuser", "default": false },
          "full_name": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Full Name"
          },
          "name": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Name" },
          "about": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "About" },
          "avatar_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Avatar Url"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" }
        },
        "type": "object",
        "required": ["email", "id"],
        "title": "UserPublic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": { "SupabaseBearer": { "type": "http", "scheme": "bearer" } }
  }
}